change validation in frontend from parseInt to parseFloat for amount

+++ frontend/src/TransactionAddUpdateDelete/TransactionAddUpdateDelete.tsx
@@ -54,7 +54,7 @@ export default function TransactionAddUpdateDelete(props: Props) {
     const handleChangeAmount = (event: React.ChangeEvent<HTMLInputElement>) => {
         props.setAmount(event.target.value)
 
-        if (parseInt(event.target.value) <= 0) {
+        if (parseFloat(event.target.value) <= 0) {
             setAmountIsError(true)
             setSaveButtonError(true)
